problem solved
Suppose you have some words and you want to right justify them, that is, align them to
the right. Create a program that reads some words and print it all right justified, in the
same order as they appear in the input.
Input:
The first line of the input will contain an integer N (1 ≤ N ≤ 50) indicating the number
of following words. Each word is composed of up to 50 letters (‘A’-‘Z’ or 'a'-'z') and
will contain at least one letter.
Output:
Print the words padded on the left with space characters so that they are all the same
length as the longest word found in that text.

Sample Input       
3
Bob
Tommy
Jim

Sample Output
      Bob
Tommy
       Jim

4
LONGEST
a
LONGER
SHORT

LONGEST
               a
  LONGER
     SHORT